deploy/classifier/vit/requirements.txt
transformers>=4.36.0
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime-gpu>=1.16.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tritonclient[all]>=2.40.0
